+++ CHANGELOG.md
@@ -172,8 +172,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Bootstrap options
 
 
-[Unreleased]: https://github.com/goph/fxt/compare/v0.6.2...HEAD
-[0.6.1]: https://github.com/goph/fxt/compare/v0.6.1...v0.6.2
+[Unreleased]: https://github.com/goph/fxt/compare/v0.6.3...HEAD
+[0.6.3]: https://github.com/goph/fxt/compare/v0.6.3...v0.6.3
+[0.6.2]: https://github.com/goph/fxt/compare/v0.6.1...v0.6.2
 [0.6.1]: https://github.com/goph/fxt/compare/v0.6.0...v0.6.1
 [0.6.0]: https://github.com/goph/fxt/compare/v0.5.2...v0.6.0
 [0.5.2]: https://github.com/goph/fxt/compare/v0.5.1...v0.5.2